+++ Assets/Scripts/System/HealthSystem.cs
@@ -2,19 +2,19 @@ using UnityEngine;
 using System;
 public class HealthSystem : MonoBehaviour
 {
-    private float _maxHealth;
-    private float _currentHealth;
+    private int _maxHealth;
+    private int _currentHealth;
 
-    public event Action<float, float> OnHealthChanged;
+    public event Action<int, int> OnHealthChanged;
     public event Action<GameObject> OnDamageTaken;
 
-    public void Initialize(float maxHealth)
+    public void Initialize(int maxHealth)
     {
         _maxHealth = maxHealth;
         _currentHealth = _maxHealth;
     }
 
-    public void TakeDamage(float damage, GameObject damageSource = null)
+    public void TakeDamage(int damage, GameObject damageSource = null)
     {
         _currentHealth = Mathf.Max(0, _currentHealth - damage);
 
@@ -24,14 +24,14 @@ public class HealthSystem : MonoBehaviour
         Debug.Log($"Damage taken: {damage} by {gameObject.name}. Current health: {_currentHealth}/{_maxHealth}");
     }
 
-    public void Heal(float heal)
+    public void Heal(int heal)
     {
         _currentHealth = Mathf.Min(_maxHealth, _currentHealth + heal);
 
         OnHealthChanged?.Invoke(_currentHealth, _maxHealth);
     }
 
-    public float CurrentHealth => _currentHealth;
-    public float MaxHealth => _maxHealth;
+    public int CurrentHealth => _currentHealth;
+    public int MaxHealth => _maxHealth;
     public bool IsAlive => _currentHealth <= 0;
 }
